Create 2 functions to check params and parse them
First function check type of parameter. If parameter has type string parse it to number. In other case return default value. In second function in parameters we have object which contains page and perPage parameters. We parse that parameters use first function and in final step we return object with this two updated(parsed) parameters